requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
docx==0.2.4
exceptiongroup==1.2.0
faiss-cpu==1.8.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.1
langsmith==0.1.27
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
yarl==1.9.4
